Update app icon and launcher label
Changed the app icon and launcher label to "Fruits Hub" for a more descriptive and visually appealing branding. Updated the necessary files in both Android and iOS platforms. Added the flutter_launcher_icons package as a dev dependency and configured it to use the new app icon. Also updated some transitive dependencies.