feat: Add Google Sign-In functionality
This commit adds the ability for users to sign in with their Google account using the `google_sign_in` package. The `FirebaseAuthService` now includes a `signInWithGoogle` method that uses the `GoogleSignIn` package to authenticate the user. The `AuthRepoImpl` now includes a `signinWithGoogle` method that returns a `UserEntity` upon successful authentication. The necessary dependencies have also been added to the `pubspec.yaml` file.