Add method to add user data to database
This commit adds a new method to the AuthRepo interface that allows for adding user data to the database. The method takes in a UserEntity object and adds it to the database. Additionally, a toMap() method is added to the UserEntity class to convert the object to a map for database storage. A Material import is also added to the AuthRepo file.