Add deleteUser method to FirebaseAuthService and handle user deletion in AuthRepoImpl
This commit adds a new method `deleteUser` to `FirebaseAuthService` that allows deleting the current user. In `AuthRepoImpl`, we handle the case where an error occurs during user creation by deleting the user if it was created before the error occurred. This ensures that we don't have any orphaned user accounts in our system.